tp: drive DebugAnnotation parsing on the proto-args work stack
Previously, ProtoToArgsParser::ParseMessage and DebugAnnotationParser::
Parse called each other when a TrackEvent's DebugAnnotation contained a
proto_value of type DebugAnnotation: the proto-args type override
synchronously invoked DebugAnnotationParser::Parse, which in turn
called back into ParseMessage. Each cycle added an ~8 KB stack frame,
which on platforms without -fstack-clash-protection could bypass the
guard page.

Add ParseDebugAnnotation as a public entry on ProtoToArgsParser; it
reuses the same iterative work stack as ParseMessage by adding two
new variant alternatives (DebugAnnotationWorkItem,
NestedValueWorkItem). DebugAnnotation -> proto_value ->
DebugAnnotation is now handled by pushing further work items onto
that stack rather than calling another driver, so depth in the input
no longer grows the C++ stack.

EnableDebugAnnotationParsing() opts a parser into routing
DebugAnnotation sub-fields to the new path; track_event_parser turns
it on and the .perfetto.protos.DebugAnnotation type override is
removed. The DebugAnnotationParser class is deleted; callers use
ProtoToArgsParser::ParseDebugAnnotation directly. A regression test
parses a 1000-deep DebugAnnotation/proto_value cycle.

DebugAnnotation tests are merged into proto_to_args_parser_unittest;
EffectiveNestedValueKey gains DCHECKs on its index/parent assumptions.
