+++ Android.bp
@@ -17810,7 +17810,6 @@ filegroup {
 filegroup {
     name: "perfetto_src_trace_processor_util_proto_to_args_parser",
     srcs: [
-        "src/trace_processor/util/debug_annotation_parser.cc",
         "src/trace_processor/util/proto_to_args_parser.cc",
     ],
 }
@@ -17930,7 +17929,6 @@ filegroup {
     name: "perfetto_src_trace_processor_util_unittests",
     srcs: [
         "src/trace_processor/util/bump_allocator_unittest.cc",
-        "src/trace_processor/util/debug_annotation_parser_unittest.cc",
         "src/trace_processor/util/galloping_search_unittest.cc",
         "src/trace_processor/util/glob_unittest.cc",
         "src/trace_processor/util/gzip_utils_unittest.cc",
